(PC-8749) Move ApiKey in core folder
