--- src/pcapi/repository/api_key_queries.py
@@ -1,5 +1,0 @@
-from pcapi.models import ApiKey
-
-
-def find_api_key_by_value(key: str) -> ApiKey:
-    return ApiKey.query.filter_by(value=key).one_or_none()
